Crate druid_shell

Platform abstraction for druid toolkit.

druid-shell is an abstraction around a given platform UI & application framework. It provides common types, which then defer to a platform-defined implementation.

Re-exports

pub use piet_common as piet;

pub use piet_common::kurbo;

Modules

application
The top-level application type.

common_util
Common functions used by the backends

error
Errors at the application shell level.

mouse
Common types for representing mouse events and state

platform
Platform specific implementations.

runloop
The main application loop.

window
Platform independent window types.

Structs

Application
The top level application object.

DruidContext

MouseEvent
The state of the mouse for a click, mouse-up, or move event.

TimerToken
A token that uniquely identifies a running timer.

WindowBuilder
A builder type for creating new windows.

WindowHandle
A handle to a platform window object.

Enums

Cursor
Mouse cursors.

Error
Error codes. At the moment, this is little more than HRESULT, but that might change.

MouseButton
An indicator of which mouse button was pressed.

Traits

WinCtx
A context supplied to most WinHandler methods.

WinHandler
App behavior, supplied by the app.

Functions

draw_to_display
Draw the item to the display

set_display_pixels
Draw the pixel colours to the display at the window from (sx,sy) to (ex,ey)

show_touch
Display the touched (X, Y) coordinates

start_display
Initialise the display controller

Type Definitions

Text
The platform text factory, reexported from piet.

WindowIdType
Windows are identified by an 8-bit ID